% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gsi.mixtureSims}
\alias{gsi.mixtureSims}
\title{Does mixture simulations from a baseline given specified mixture proportions}
\usage{
gsi.mixtureSims(B, pops.f, props, N, Reps, fixed = TRUE,
  base_file_name = "mix_sim_base.txt", repu.f = NULL, samp_unit = "genes")
}
\arguments{
\item{props}{Vector of mixing proprtions parallel to the levels of pops.
These are the true mixing proportions that you want to simulate from}

\item{N}{the sample size of each sample you want to simulate. It will be approximate due to
rounding.  I could fix that eventually, but it is a bit of a hassle at the moment.}

\item{Reps}{Number of simulated samples you want to simulate}

\item{fixed}{Logical.  True means use the --fixed-Pi version (so that the exact
same number of fish are simulated each time.  False means that the origins of the fish
in the simulated sample are drawn from the props (by using the --Pi option of gsi_sim))}

\item{base_file_name}{The name for baseline file that will be created and passed to gsi_sim}

\item{repu.f}{A factor vector that is parallel to the levels of pops.f, each entry giving the
reporting unit that that particular level of pops.f belongs to.  This is used to aggregate the
estimated mixing proportions over reporting groups.  If NULL then it doesn't get done.}

\item{samp_unit}{What sampling unit to resample? Corresponds to gsi_sim's --samp_unit option.
Allowable values are "genes", "loci", and "multilocus".}

\item{B}{a genetic baseline in \code{\link{gPipe.data.frame}} format}

\item{pops.f}{a factor vector of populations each individual in B shall belong to.  levels of the
factor should be in the order that these should appear in.  pops.f must have length
equal to the number of rows of B}
}
\description{
The user can specify a gpiper input file and a vector of populations
as well as a vector of proportions of populations and a size of sample
and this will do all the leg-work do simulations with gsi_sim using
either the fixed-Pi or the Pi variants.  Currently it only extracts the
mixing proportion estimates out of it.
}

