% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gsi_aggScoresByRepUnit}
\alias{gsi_aggScoresByRepUnit}
\title{sum gsi_sim posteriors over populations in reporting units}
\usage{
gsi_aggScoresByRepUnit(x, pops.str, rep.units)
}
\arguments{
\item{x}{data frame like that obtained by reading in "pop_pofz_full_em_mle.txt" or a similar output file
from gsi_sim (with rownames=1)}

\item{pops.str}{names of the populations that are the column names of x}

\item{rep.units}{factor vector of same length as pops.str which given the reporting units of the populations in pops.str}
}
\description{
given a data frame x that has columns represententing scores (like posterior
probs) for each population in the pops.str (which should be a character
vector of population names) which are themselves grouped into reporting
units according to rep.units (preferably a factor), this returns a new data frame
with all the non-pop columns first and then the pop columns aggregated by summing
them up.
}

